requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
selenium
sqlalchemy
psycopg2-binary
